Change image type to png and add various sections to index.html
This commit updates the link for the favicon to use a png image and adds several new sections to the index.html file. These new sections include a "Where am I" section with contact information, a "Learn html" section, an ordered list of things to learn, an unordered list of technologies, and two description lists for SRM-College and front end Developer.